# finvera-be/src/main/java/com/minhnb/finvera_be/alert/controller/AlertController.java
package com.minhnb.finvera_be.alert.controller;import static com.minhnb.finvera_be.alert.dto.AlertDtos.*;import com.minhnb.finvera_be.alert.service.AlertManagementService;import jakarta.validation.Valid;import java.net.URI;import java.util.UUID;import org.springframework.http.ResponseEntity;import org.springframework.web.bind.annotation.*;
@RestController@RequestMapping("/api/v1/alerts")public class AlertController{private final AlertManagementService service;public AlertController(AlertManagementService s){service=s;}@PostMapping public ResponseEntity<AlertResponse> create(@Valid@RequestBody CreateAlertRequest r){var x=service.create(r);return ResponseEntity.created(URI.create("/api/v1/alerts/"+x.id())).body(x);}@GetMapping public PageResponse<AlertResponse> list(@RequestParam(defaultValue="20")int limit,@RequestParam(defaultValue="0")int offset){bounds(limit,offset);return service.list(limit,offset);}@GetMapping("/{id}")public AlertResponse get(@PathVariable UUID id){return service.get(id);}@PutMapping("/{id}/state")public AlertResponse state(@PathVariable UUID id,@RequestBody SetAlertStateRequest r){return service.state(id,r.enabled());}@DeleteMapping("/{id}")public ResponseEntity<Void> delete(@PathVariable UUID id){service.delete(id);return ResponseEntity.noContent().build();}@GetMapping("/{id}/evaluations")public PageResponse<EvaluationResponse> evaluations(@PathVariable UUID id,@RequestParam(defaultValue="20")int limit,@RequestParam(defaultValue="0")int offset){bounds(limit,offset);return service.evaluations(id,limit,offset);}private static void bounds(int l,int o){if(l<1||l>100||o<0||o>10000)throw new IllegalArgumentException("PAGE_BOUNDS_INVALID");}}