# finvera-be/src/main/java/com/minhnb/finvera_be/alert/controller/NotificationController.java
package com.minhnb.finvera_be.alert.controller;import static com.minhnb.finvera_be.alert.dto.AlertDtos.*;import com.minhnb.finvera_be.alert.service.NotificationService;import java.util.UUID;import org.springframework.web.bind.annotation.*;
@RestController@RequestMapping("/api/v1/notifications")public class NotificationController{private final NotificationService service;public NotificationController(NotificationService s){service=s;}@GetMapping public PageResponse<NotificationResponse> list(@RequestParam(defaultValue="false")boolean unreadOnly,@RequestParam(defaultValue="20")int limit,@RequestParam(defaultValue="0")int offset){if(limit<1||limit>100||offset<0||offset>10000)throw new IllegalArgumentException("PAGE_BOUNDS_INVALID");return service.list(unreadOnly,limit,offset);}@GetMapping("/unread-count")public UnreadCountResponse unread(){return new UnreadCountResponse(service.unread());}@GetMapping("/{id}")public NotificationResponse get(@PathVariable UUID id){return service.get(id);}@PutMapping("/{id}/read")public NotificationResponse read(@PathVariable UUID id){return service.read(id);}@PutMapping("/read-all")public ReadAllResponse readAll(){return service.readAll();}}